Update README.md
This genetic algorithm attempts to take a larger sample of subjects and get as close of a match to a separate smaller sample based on one or more 
numerical attributes that are collected between both groups. The fitness of the match is based on euclidean distance between the shared attributes of the samples. 
Random mutations are introduced in attempt to get out of local minima and converge to a true solution.